♻️ 型周りをリファクタ
